video_pipeline/requirements.txt
# Core LLM
anthropic>=0.40.0
openai>=1.50.0
python-dotenv>=1.0.0

# Image generation
torch>=2.1.0
diffusers>=0.27.0
transformers>=4.40.0
accelerate>=0.30.0
safetensors>=0.4.0
xformers; platform_system != "Darwin"

# Audio / TTS
TTS>=0.22.0
elevenlabs>=1.5.0

# Video processing
moviepy==1.0.3
imageio>=2.34.0
imageio-ffmpeg>=0.4.9

# Utilities
Pillow>=10.0.0
numpy>=1.24.0
pydub>=0.25.0

# Optional UI
streamlit>=1.32.0
